router: Share blink method across children
`TurrisRouter` `blink` method code is superset of `OmniaRouter` `blink`
method code. The only difference is list of leds and in translating the
`lanX` device name to `lan` when setting a color.

Children of `Router` need to set `self.leds` list and then call super
class constructor as well to set the (shared) config.